Changes to modernize and clarify
delete - removed "shelley" sub-command as it is deprecated.
add - ledger era was wrong so added  "--allegra-era \ " where needed to fix mismatch error. 
revision - changed the wallet amount to start with the current daily faucet flow limit of 1000 ada (1,000,000,000 lovelace) and corrected calculations to reflect correctly.
revision - moved instructions to get the index of the TxHash to the "Draft the transaction" section as this is the first place it is needed and referred to it in the "Build the transaction" section where additional clarification was also added.