current software versions; minor cosmetic edits
COMMENTS:

1) The {% hint style="info" %} was not working for me, so I removed it (maybe unnecessary to do this?);

2) In my opinion the introductory bit about Virtual Machines, AWS and System requirements needs amending. I wrote this in the forum: 

"Information to be revised:
“An x86 host (AMD or Intel), Virtual Machine or AWS instance with at least 2 cores, 4GB of RAM and at least 10GB of free disk space”

Suggested revisions:
The host does not need to be a VM or specifically an AWS instance. Suggesting only one cloud-computing provider is also a bit restrictive: there are many others and, if compiling a guide for each one is impractical, the suggestion should be limited to “a cloud-computing provider”. I realise there is a section in SPS (https://cardano-foundation.gitbook.io/stake-pool-course/lessons/lesson-1/alternative-to-aws 1) which points to an alternative, this is why it should be reworded here. Also, nodes can be hosted on local machines, as long as the operating system is correct.

Further, the system requirements are in contrast with those stated in other guides (for example, Coincashew, https://www.coincashew.com/coins/overview-ada/guide-how-to-build-a-haskell-stakepool-node#minimum-stake-pool-hardware-requirements 1 ) and also the official github repository https://github.com/input-output-hk/cardano-node/releases/tag/1.24.2 1. These system requirements need to be made clear and homogeneous in all sources."

I am unsure how to word these changes, and welcome input from more expert people :) 

3) I have not experienced the need to delete the db folder (seems a bit extreme). Does anyone else feel like this is a bit overkill? 

4) Are we still aiming for this guide to be exclusive to Testnet?

Cheers,

A