removed Shelley references
The infographic at the bottom will need amending too. Does anyone have access to the original file? Was it vector?

Also, should there be a note showing what happens if you do run the command with shelley subcommand? This still works, but will soon be removed. 

Looking forward to your thoughts,

A